requirements.txt
asgiref==3.11.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dj-database-url==3.0.1
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
httplib2==0.31.2
idna==3.11
packaging==26.0
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.3.2
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
whitenoise==6.11.0
